feat(ips): update guide links
ref: #MANAGER-17644
