+++ packages/manager/apps/ips/public/translations/onboarding/Messages_fr_FR.json
@@ -8,12 +8,14 @@
   "orderButtonLabel": "Commander une IP",
   "byoipButtonLabel": "Bring your own IP",
   "guideCategory": "Tutoriel",
-  "guide1Title": "Déplacer une Additional IP",
-  "guide1Description": "Découvrez comment déplacer une Additional IP depuis l'espace client ou via les API OVHcloud",
-  "guide2Title": "Configurer un bloc d'IP dans le vRack",
-  "guide2Description": "Découvrez comment configurer un bloc d'adresses IP publiques dans le vRack",
-  "guide3Title": "Ajouter un bloc IP",
-  "guide3Description": "Commander un bloc IP sur votre Hosted Private Cloud",
+  "guide1Title": "Configurer une Additional IP",
+  "guide1Description": "Découvrez comment ajouter des adresses Additional IP à la configuration de votre instance",
+  "guide2Title": "Configurer son adresse IP en alias",
+  "guide2Description": "Découvrez comment ajouter des Additional IP à votre configuration",
+  "guide3Title": "Configurer le réseau sur Proxmox VE",
+  "guide3Description": "Découvrez comment configurer le réseau sur Proxmox VE",
+  "guide4Title": "Concepts - Additional IP ou Floating IP sur Public Cloud",
+  "guide4Description": "Découvrez comment utiliser les adresses IP publiques dans le cadre du réseau Public Cloud",
 
   "advantagesTitle": "Avantages des Additional IP",
   "advantage1Title": "Capacités de fail-over",
