+++ packages/manager/apps/ips/src/utils/index.ts
@@ -2,5 +2,5 @@ export * from './iamActions.constants';
 export * from './ipFormatter';
 export * from './getTypeByServiceName';
 export * from './getLinkByServiceName';
-export * from './links.constants';
 export * from './ipFeatureAvailbility';
+export * from './useGuideUtils';
